Add additional logging to the heartbeat at trace level for future diagnostics of what the heartbeat is doing. Catch any exception when the heartbeat tries to connect, log it, and proceed allowing the heartbeat to be rescheduled. An example of one of these exceptions is being unable to resolve a DNS name due to network connectivity issues thrown from the TcpResourceAddressFactorySpi. When this exception bubbles up to the heartbeat the heartbeat was not getting rescheduled. That is now fixed such that when network connectivity is restored the heartbeat correctly re-binds the service in question.
Ensure the address used in the WsebConnector is the originally build address with all the correct transports and options.  This corrects an issue where WSEB over reverse SOCKS was removing the transport, not noticing it should bind rather than connect, and then ending up connecting to the wrong place rather than negotiating a reverse SOCKS connection.  Since standard behavior is to connect to the address giving, preserve that behavior.